+++ printf/libft/ft_recursive_power.c
@@ -6,14 +6,14 @@
 /*   By: loiberti <marvin@le-101.fr>                +:+   +:    +:    +:+     */
 /*                                                 #+#   #+    #+    #+#      */
 /*   Created: 2018/10/08 19:03:12 by loiberti     #+#   ##    ##    #+#       */
-/*   Updated: 2018/10/08 20:00:11 by loiberti    ###    #+. /#+    ###.fr     */
+/*   Updated: 2018/11/26 16:31:56 by loiberti    ###    #+. /#+    ###.fr     */
 /*                                                         /                  */
 /*                                                        /                   */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_recursive_power(int nb, int power)
+size_t	ft_recursive_power(int nb, int power)
 {
 	if (!power)
 		return (1);
